requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pymongo==4.3.3
python-dotenv==1.0.0
PyJWT==2.6.0
bcrypt==4.0.1